@@ -2,9 +2,9 @@
 
 	<PropertyGroup>
 		<OutputType>Library</OutputType>
-		<TargetFramework>net472</TargetFramework>
+		<TargetFramework>net48</TargetFramework>
 		<Configurations>Debug;Release</Configurations>
-		<LangVersion>9.0</LangVersion>
+		<LangVersion>12.0</LangVersion>
 
 
 		<!--The file where mod information which is required for publishing mod on PDX mods are stored-->
@@ -14,12 +14,10 @@
 	<!--Imports must be after PropertyGroup block-->
 	<Import Project="$([System.Environment]::GetEnvironmentVariable('CSII_TOOLPATH', 'EnvironmentVariableTarget.User'))\Mod.props" />
 	<Import Project="$([System.Environment]::GetEnvironmentVariable('CSII_TOOLPATH', 'EnvironmentVariableTarget.User'))\Mod.targets" />
-	<ItemGroup>
-	  <Compile Remove="Components\OtherComponents.cs" />
-	</ItemGroup>
 
 	<ItemGroup>
 		<Reference Include="Colossal.Mathematics">
+			<Private>false</Private>
 		</Reference>
 		<Reference Include="Game">
 			<Private>false</Private>
